requirements.txt
aiogram==3.4.1
aiohttp
wheel
aiosignal==1.3.1
argon2-cffi==21.1.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
boto3==1.18.63
botocore==1.21.65
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-storages==1.12.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
jmespath==0.10.0
magic-filter==1.0.12
multidict==6.0.4
mutagen==1.45.1
oauthlib==3.2.2
Pillow
psycopg2-binary
psycopg
pycparser==2.21
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
whitenoise==5.3.0
yarl==1.9.3
django-storages 
environs
python-dotenv
django-compressor
dj-database-url
